Fix #1604 - FilePlayer does not work if I passed an array of urls (#1612)
PR #1538 was setting the src of the video element to undefined. Because
of this, even if the users passed an array of urls, the video element
would not play becuase it got a `src` of `undefined`;

After fixing the undefined `src` issue, another PR was making a infinite
loop issue when user passed an array of urls. There, the author set the
`srcObject` to null. There as well, a check for `url` prop was missing.

This PR aims to fix both of these overlooked aspects.